Getting rid of the background
At this stage, you need to get rid of the difference noise of the brightness of the pixels. Find the level of average values and lower the brightness of all pixels below it.
To do this, a row of pixels is selected from the middle of the image. Next is the most common value of all the items in the list, and is multiplied by 25%. In code, this is done as follows.