Isolation of a character vertically.
In the algorithm, the row variable, called the "counting line", is equal to the number of lines whose sum of brightnesses is more than five hundred. If after successive lines of sums there is a line with the sum of less than five hundred units and the counting line is less than a hundred, then all lines to the counting line are deleted and the counting line is reset. If the counting line is more than a hundred, then all lines after the counting line are deleted.