Isolation of a character horizontally.
In the algorithm for selecting a character horizontally, as well as in the algorithm for selecting a row vertically, there is a counting column responsible for the number of consecutive columns with a sum of more than three hundred units. In the case of a sum of less than three hundred units and a counting column of less than fifteen, all left columns are deleted. And in the case of a countable column of more than fifteen, the third, fourth, fifth and sixth stages of image fragmentation are performed.