Isolation of a character vertically.
The algorithm consists of four variables: row - the number of lines the sum of which exceeds five hundred units, high - storing the beginning of the term for deletion, depth - a variable for passing over the width of the character, symbol - storing the values of the symbol divided into parts. The algorithm is to get rid of lines whose sum is less than five hundred units, as well as objects whose size is less than 10 lines. Also, the algorithm is able to determine a divided object into several parts, such as dividing “:”.