Calculation of initial values.
At the beginning of the algorithm, the maximum axis of the image is selected, along which the magnification factor will be found.
After that, the magnification factor is multiplied by 28 and a new image size is found, which creates an empty picture.
To write the input image to the created one, there are not enough coordinates, when recording along which the input image was in the middle of the created one. Finding these coordinates is as follows.
Half of the side of the created image is subtracted half of the input side.