An exception
On the first cycle, arithmetic characters are excluded at the beginning of the list, except for subtraction.
In the next cycle, all arithmetic signs from the last position are excluded.
In the beam, if an error occurs on the index, it means that the input list was empty initially. Therefore, you need to display a message about the absence of an example and close the function.