Conversion
list_element - a list to record the result,
math_sig - a list with the characters used,
number is an intermediate variable.
The cycle goes through the training_list input list element by element. If the input value is a digit, then it deletes itself by 0.1, after which the variable number is added and all this is multiplied by ten. Thus, from the numbers you can make a number. If after this comes the arithmetic sign, then the number and sign are written to the end of the list_element list. This algorithm continues until the entire training_list ends.
In the end, all unnecessary variables are erased.