Getting rid of extra characters
If suddenly the user entered an incorrect sequence or extra characters.
Or the neural network could not perceive the graphic record, then some numbers could disappear, forming a sequential arrangement of characters.
Such a decision record will form an error and the program will stop.
Therefore, an algorithm is required to prevent this error.
The algorithm consists of three conditions:
if the iteration element is a minus;
if it is a different sign, and this is addition, multiplication and division; if the item is a number.
The program also goes through the entire list element by element.
Extra characters are removed due to error conditions.
In cases of waiting for the next or previous element with a value of a number, the transition to the next one is performed.
But if the number is absent, and the arithmetic sign is present, then the requested sign is deleted.
When the iteration sign is minus, and the next element is a number, then this number changes the sign to negative.
If the previous value is not a number, but an arithmetic sign, then the minus sign is deleted.
And if this is a number, then the minus sign is replaced by a plus.
As a result, the algorithm erases the extra arithmetic signs, leaving only the final ones closest to the number, and gets rid of the minus sign, replacing it with a plus, and the next value multiplies by minus one, inverting its sign.