Counting equation
At this point, an example will be counted for three arithmetic operations: multiplication, division, and addition.
For the correct calculation of the expression, it is necessary to observe the rules of priority of the sign, that is, first multiplication or division is performed, then addition.
Since the sequence of division and multiplication can be any, it is necessary to determine which action will begin first.
Three variables are created:
addition_address (addition address),
multiplication_address (multiplication address),
division_address (division address) and
two latch trigger variables multiplication_kay (True multiplication is present),
division_kay (False division is present).
The algorithm works according to the method of traversing the entire list using a loop.
First, the program remembers the leftmost addresses of multiplication and division, then compares and selects the action with the lowest address, and then calculates the numbers around it.
If there are no signs of multiplication and division, locking triggers are triggered, which deactivates the conditions of these signs and activates the condition of the addition sign.
After that, all additions with positive or negative numbers are performed.
When performing any action (with two numbers), its result is written to the first number, and the arithmetic sign and the second number are removed from the list.
This sequence continues until only the answer remains.
As a result of the whole algorithm, an answer is added to the duplicated list.
Further, all elements of the extended list are converted into a character data type.
After that, the entire list is converted into one line of characters, which will be displayed on the user's screen.